+++ client/src/shared/lib/models/models.ts
@@ -49,5 +49,6 @@ export interface RegisterTerminal {
 export enum StatusCodes {
   OK = 200,
   CREATED = 201,
-  BAD_REQUEST = 400
+  BAD_REQUEST = 400,
+  NOT_FOUND = 404,
 }
